requirements.txt
Flask
gunicorn
Pillow
ImageHash
